requirements.txt
playwright
psycopg2-binary
python-telegram-bot